[minor] remove `BrowserKitBrowser`, add `Browser::client()`
