+++ phpstan-baseline.neon
@@ -11,21 +11,6 @@ parameters:
 			path: src/Browser.php
 
 		-
-			message: "#^Method Zenstruck\\\\Browser\\:\\:selectField\\(\\) has parameter \\$value with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\:\\:selectFieldOptions\\(\\) has parameter \\$values with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser.php
-
-		-
-			message: "#^Parameter \\#2 \\$path of method Behat\\\\Mink\\\\Element\\\\TraversableElement\\:\\:attachFileToField\\(\\) expects string, array\\<string\\>\\|string given\\.$#"
-			count: 1
-			path: src/Browser.php
-
-		-
 			message: "#^Method Zenstruck\\\\Browser\\\\Assertion\\\\SameUrlAssertion\\:\\:__construct\\(\\) has parameter \\$partsToMatch with no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Browser/Assertion/SameUrlAssertion.php
@@ -54,41 +39,6 @@ parameters:
 			message: "#^Property Zenstruck\\\\Browser\\\\Assertion\\\\SameUrlAssertion\\:\\:\\$partsToMatch type has no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Browser/Assertion/SameUrlAssertion.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:delete\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:get\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:patch\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:post\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:put\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:request\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
-
-		-
-			message: "#^Method Zenstruck\\\\Browser\\\\BrowserKitBrowser\\:\\:setDefaultHttpOptions\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Browser/BrowserKitBrowser.php
 
 		-
 			message: "#^Cannot access offset 'query' on array\\{scheme\\?\\: string, host\\?\\: string, port\\?\\: int, user\\?\\: string, pass\\?\\: string, path\\?\\: string, query\\?\\: string, fragment\\?\\: string\\}\\|false\\.$#"
@@ -166,6 +116,56 @@ parameters:
 			path: src/Browser/HttpOptions.php
 
 		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:delete\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:get\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:patch\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:post\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:put\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:request\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:selectField\\(\\) has parameter \\$value with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:selectFieldOptions\\(\\) has parameter \\$values with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\KernelBrowser\\:\\:setDefaultHttpOptions\\(\\) has parameter \\$options with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
+			message: "#^Parameter \\#2 \\$path of method Behat\\\\Mink\\\\Element\\\\TraversableElement\\:\\:attachFileToField\\(\\) expects string, array\\<string\\>\\|string given\\.$#"
+			count: 1
+			path: src/Browser/KernelBrowser.php
+
+		-
 			message: "#^Method Zenstruck\\\\Browser\\\\Mink\\\\BrowserKitDriver\\:\\:getFormField\\(\\) should return Symfony\\\\Component\\\\DomCrawler\\\\Field\\\\FormField but returns array\\<Symfony\\\\Component\\\\DomCrawler\\\\Field\\\\FormField\\>\\|Symfony\\\\Component\\\\DomCrawler\\\\Field\\\\FormField\\.$#"
 			count: 1
 			path: src/Browser/Mink/BrowserKitDriver.php
@@ -214,4 +214,19 @@ parameters:
 			message: "#^Parameter \\#3 \\.\\.\\.\\$values of function sprintf expects bool\\|float\\|int\\|string\\|null, array\\|bool\\|string given\\.$#"
 			count: 1
 			path: src/Browser/Mink/PantherDriver.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\PantherBrowser\\:\\:selectField\\(\\) has parameter \\$value with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/PantherBrowser.php
+
+		-
+			message: "#^Method Zenstruck\\\\Browser\\\\PantherBrowser\\:\\:selectFieldOptions\\(\\) has parameter \\$values with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Browser/PantherBrowser.php
+
+		-
+			message: "#^Parameter \\#2 \\$path of method Behat\\\\Mink\\\\Element\\\\TraversableElement\\:\\:attachFileToField\\(\\) expects string, array\\<string\\>\\|string given\\.$#"
+			count: 1
+			path: src/Browser/PantherBrowser.php
